Temporarily factor out MockResponse from pytest dependent file
